#9 ログアウトボタンをマイページに移動する
