+++ Tumolink2iOSClient/Controller/HomeVC.swift
@@ -12,7 +12,6 @@ import Firebase
 class HomeVC: UIViewController {
     
     // MARK: Outlets
-    @IBOutlet weak var loginBtn: UIBarButtonItem!
     @IBOutlet weak var tableView: UITableView!
     @IBOutlet weak var activityIndicator: UIActivityIndicatorView!
     
@@ -25,13 +24,12 @@ class HomeVC: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        setupInitialAnonymouseUser()
         db = Firestore.firestore()
+        setupInitialAnonymouseUser()
         setupTableView()
     }
     
     override func viewDidAppear(_ animated: Bool) {
-        setupLoginBtn()
         setupMySpots()
     }
     
@@ -58,27 +56,6 @@ class HomeVC: UIViewController {
                 }
             }
         }
-    }
-    
-    private func setupLoginBtn() {
-        if let user = Auth.auth().currentUser {
-            print("login as \(user.uid)")
-        } else {
-            print("currentUser is nil")
-        }
-        
-        if let user = Auth.auth().currentUser , !user.isAnonymous {
-            loginBtn.title = "ログアウト"
-        }  else {
-            loginBtn.title = "ログイン"
-        }
-    }
-    
-    // ログインフローに遷移するためのメソッド
-    fileprivate func presentLoginController() {
-        let storyboard = UIStoryboard(name: Storyboard.LoginStoryboard, bundle: nil)
-        let controller = storyboard.instantiateViewController(withIdentifier: StoryboardId.LoginVC)
-        present(controller, animated: true, completion: nil)
     }
     
     // テーブルに表示されるセルのデータを制御するメソッド
@@ -114,27 +91,6 @@ class HomeVC: UIViewController {
     }
     
     // MARK: Actions
-    @IBAction func loginClicked(_ sender: Any) {
-        guard let user = Auth.auth().currentUser else { return }
-        if user.isAnonymous {
-            presentLoginController()
-        } else {
-            do {
-                try Auth.auth().signOut()
-                UserService.logoutUser()
-                Auth.auth().signInAnonymously { (result, error) in
-                    if let error = error {
-                        debugPrint(error)
-                        Auth.auth().handleFireAuthError(error: error, vc: self)
-                    }
-                    self.presentLoginController()
-                }
-            } catch {
-                debugPrint(error)
-                Auth.auth().handleFireAuthError(error: error, vc: self)
-            }
-        }
-    }
 }
 
 extension HomeVC : UITableViewDelegate, UITableViewDataSource {
