+++ Tumolink2iOSClient/Controller/MyPageVC.swift
@@ -8,12 +8,14 @@
 
 import UIKit
 import Kingfisher
+import FirebaseAuth
 
 class MyPageVC: UIViewController {
 
     // MARK: Outlets
     @IBOutlet weak var usernameTxt: UILabel!
     @IBOutlet weak var profileImg: UIImageView!
+    @IBOutlet weak var logoutBtn: FullRoundedButton!
     
     // MARK: Functions
     override func viewDidLoad() {
@@ -22,6 +24,7 @@ class MyPageVC: UIViewController {
     }
     
     override func viewDidAppear(_ animated: Bool) {
+        setupLoginBtn()
         if let username = usernameTxt.text {
             if username != UserService.user.username {
                 setupProfile()
@@ -44,4 +47,43 @@ class MyPageVC: UIViewController {
             }
         }
     }
+    
+    private func setupLoginBtn() {
+        if let user = Auth.auth().currentUser , !user.isAnonymous {
+            logoutBtn.setTitle("ログアウト", for: .normal)
+        }  else {
+            logoutBtn.setTitle("ログイン", for: .normal)
+        }
+    }
+    
+    // ログインフローに遷移するためのメソッド
+    fileprivate func presentLoginController() {
+        let storyboard = UIStoryboard(name: Storyboard.LoginStoryboard, bundle: nil)
+        let controller = storyboard.instantiateViewController(withIdentifier: StoryboardId.LoginVC)
+        present(controller, animated: true, completion: nil)
+    }
+    
+    // MARK: Actions
+    @IBAction func logoutClicked(_ sender: Any) {
+        guard let user = Auth.auth().currentUser else { return }
+        if user.isAnonymous {
+            presentLoginController()
+        } else {
+            do {
+                try Auth.auth().signOut()
+                UserService.logoutUser()
+                Auth.auth().signInAnonymously { (result, error) in
+                    if let error = error {
+                        debugPrint(error)
+                        Auth.auth().handleFireAuthError(error: error, vc: self)
+                    }
+                    self.presentLoginController()
+                }
+            } catch {
+                debugPrint(error)
+                Auth.auth().handleFireAuthError(error: error, vc: self)
+            }
+        }
+    }
+    
 }
